Moved font size setting into body, added border color, and changed font size
I set the font size in the HTML selector which was affecting the sizing of
everything else because I'm using REM units. I added a border colors that will
be used in both the header and the planet component. I changed the font sizes
to fit the design after moving the font size setting to the body selector.