Added a spinner for when the planet images are transitioning
I tested the app on a throttled connection and the component's text was changing
before the planet image. I added a loading spinner, some load state, and an onLoad
event handler to remove the spinner when the images are loaded. This ensures
the images aren't lagging behind the rest of the component and gives users on
slower connections a more fluid experience.